Add scala.js support for enumeratum
