+++ modules/enumeratum/shared/src/main/scala/enumeratum/values/Ciris.scala
@@ -8,7 +8,7 @@ package enumeratum.values
 
 import cats.Show
 import ciris.ConfigDecoder
-import org.tpolecat.typename.TypeName
+import enumeratum.internal.TypeName
 
 object Ciris {
   final def enumConfigDecoder[ValueType, EntryType <: ValueEnumEntry[ValueType]](
@@ -18,5 +18,5 @@ object Ciris {
     typeName: TypeName[EntryType],
     show: Show[ValueType]
   ): ConfigDecoder[String, EntryType] =
-    decoder.mapOption(typeName.value)(enum.withValueOpt)
+    decoder.mapOption(typeName.value)(`enum`.withValueOpt)
 }
